pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tamx"
version = "0.1.0"
description = "Token Activation Maps (TAM) for Multimodal Large Language Models"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "T.J. Wei"}
]
keywords = ["mllm", "interpretability", "visualization", "vlm", "qwen-vl"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.9.1",
    "transformers>=5.0.0",
    "numpy>=1.24.0",
    "scipy>=1.17.0",
    "safetensors>=0.4.0",
    "Pillow>=12.1.0",
    "qwen-vl-utils>=0.0.14",
    "torchcodec>=0.9.0",
    "torchvision>=0.24.1",
    "accelerate>=1.12.0"
]

[project.optional-dependencies]
vis = [
    "fastapi>=0.128.0",
    "pydantic>=2.10.0",
    "uvicorn>=0.40.0",
]
all = [
    "tamx[vis]",
]

[project.urls]
Homepage = "https://github.com/joinn99/tamx"
Repository = "https://github.com/joinn99/tamx"

[tool.setuptools.packages.find]
include = ["tamx*", "webapp*"]

[tool.black]
line-length = 100
target-version = ["py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "W", "F", "I", "B", "C4", "UP"]
